services/authentication/requirements.txt
fastapi
uvicorn
pydantic
email-validator
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-multipart
